requirements.txt
django==3.0.8
django-extensions==2.2.9
six==1.14.0
django-richtextfield==1.5.0
django-crispy-forms==1.8.1
Pillow==7.2.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
